Create workingWithLogFiles.py
Imagine one of your colleagues is struggling with a program that keeps throwing an error. Unfortunately, the program's source code is too complicated to easily find the error there. The good news is that the program outputs a log file you can read! Let's write a script to search the log file for the exact error, then output that error into a separate file so you can work out what's wrong.

What you'll do
Write a script to search the log file using regex to find for the exact error.
Report the error into a separate file so you know what's wrong for further analysis.
You'll have 90 minutes to complete this lab.